// src/services/models/response-result-recommend-danger-solution-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SeWork API
 * SeWork API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { RecommendDangerSolutionResponse } from './recommend-danger-solution-response';

/**
 * 
 * @export
 * @interface ResponseResultRecommendDangerSolutionResponse
 */
export interface ResponseResultRecommendDangerSolutionResponse {
    /**
     * 
     * @type {string}
     * @memberof ResponseResultRecommendDangerSolutionResponse
     */
    'code'?: string;
    /**
     * 
     * @type {RecommendDangerSolutionResponse}
     * @memberof ResponseResultRecommendDangerSolutionResponse
     */
    'data'?: RecommendDangerSolutionResponse;
    /**
     * 
     * @type {string}
     * @memberof ResponseResultRecommendDangerSolutionResponse
     */
    'message'?: string;
}