// src/services/models/response-result-upsert-company-danger-solution-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SeWork API
 * SeWork API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import { UpsertCompanyDangerSolutionResponse } from './upsert-company-danger-solution-response';

/**
 * 
 * @export
 * @interface ResponseResultUpsertCompanyDangerSolutionResponse
 */
export interface ResponseResultUpsertCompanyDangerSolutionResponse {
    /**
     * 
     * @type {string}
     * @memberof ResponseResultUpsertCompanyDangerSolutionResponse
     */
    'code'?: string;
    /**
     * 
     * @type {UpsertCompanyDangerSolutionResponse}
     * @memberof ResponseResultUpsertCompanyDangerSolutionResponse
     */
    'data'?: UpsertCompanyDangerSolutionResponse;
    /**
     * 
     * @type {string}
     * @memberof ResponseResultUpsertCompanyDangerSolutionResponse
     */
    'message'?: string;
}